+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * `add_sparklines()` creates an HTML widget to display variable importance scores with a sparkline representation of each features effect (i.e., its partial dependence function) [(#64)](https://github.com/koalaverse/vip/issues/64).
 
+* New method `vi_model.nnet()` which support both Olden and Garson algorithms for neural networks fit using the __nnet__ package.
+
 ## Breaking changes
 
 * The `pred_fun` argument in `vi_permute()` has been changed to `pred_wrapper`.